Dev

See merge request cfs/fnsc!194

+++ service/src/main/java/ir/demisco/cfs/service/impl/DefaultFinancialSecurityUser.java
@@ -3,9 +3,13 @@ package ir.demisco.cfs.service.impl;
 import ir.demisco.cfs.model.dto.request.FinancialUsersInputModelRequest;
 import ir.demisco.cfs.model.dto.response.FinancialSecurityUserResponse;
 import ir.demisco.cfs.model.entity.FinancialUser;
+import ir.demisco.cfs.service.api.FinancialAlternativeUsersService;
+import ir.demisco.cfs.service.api.FinancialSecPermissionScopeService;
 import ir.demisco.cfs.service.api.FinancialSecurityUserService;
+import ir.demisco.cfs.service.api.UserPermissionService;
 import ir.demisco.cfs.service.repository.ApplicationUserRepository;
 import ir.demisco.cfs.service.repository.FinancialUserRepository;
+import ir.demisco.cloud.core.middle.exception.RuleException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,10 +20,18 @@ import java.util.stream.Collectors;
 public class DefaultFinancialSecurityUser implements FinancialSecurityUserService {
     private final ApplicationUserRepository applicationUserRepository;
     private final FinancialUserRepository financialUserRepository;
+    private final FinancialAlternativeUsersService financialAlternativeUsersService;
+    private final UserPermissionService userPermissionService;
+    private final FinancialSecPermissionScopeService financialSecPermissionScopeService;
 
-    public DefaultFinancialSecurityUser(ApplicationUserRepository applicationUserRepository, FinancialUserRepository financialUserRepository) {
+    public DefaultFinancialSecurityUser(ApplicationUserRepository applicationUserRepository, FinancialUserRepository financialUserRepository,
+                                        FinancialAlternativeUsersService financialAlternativeUsersService, UserPermissionService userPermissionService,
+                                        FinancialSecPermissionScopeService financialSecPermissionScopeService) {
         this.applicationUserRepository = applicationUserRepository;
         this.financialUserRepository = financialUserRepository;
+        this.financialAlternativeUsersService = financialAlternativeUsersService;
+        this.userPermissionService = userPermissionService;
+        this.financialSecPermissionScopeService = financialSecPermissionScopeService;
     }
 
 
@@ -53,6 +65,18 @@ public class DefaultFinancialSecurityUser implements FinancialSecurityUserServic
 
     @Override
     public Boolean deleteFinancialUser(Long financialUserId) {
+        Long financialUserAlternativeByFinancialUserId = financialAlternativeUsersService.getFinancialUserAlternativeByFinancialUserId(financialUserId);
+        if (financialUserAlternativeByFinancialUserId > 0L) {
+            throw new RuleException("امکان حذف این کاربر مالی وجود ندارد.");
+        }
+        Long userPermissionByFinancialUserId = userPermissionService.getUserPermissionByFinancialUserId(financialUserId);
+        if (userPermissionByFinancialUserId > 0L) {
+            throw new RuleException("امکان حذف این کاربر مالی وجود ندارد.");
+        }
+        Long userPermissionScopeByFinancialUserId = financialSecPermissionScopeService.findUserPermissionScopeByFinancialUserId(financialUserId);
+        if (userPermissionScopeByFinancialUserId > 0L) {
+            throw new RuleException("امکان حذف این کاربر مالی وجود ندارد.");
+        }
         financialUserRepository.deleteById(financialUserId);
         return true;
     }
